[wificontrol] Hides assertion triggering.
It seems details request may be sent multiple times
because requirest's data is implicitly cleared by SendData()
and that is interpreted as a new request could be sent.
This code hides the issue by avoiding ASSERT() triggering
in Submit but ideally the data should be cleared only once
the request is done. Moreover using a single request also makes
a lot of BSS_ADDED events is ignored. Maybe it should be changed
to use a multiple requests (*shrug*)?